Handle multiple slide boxes
This fixes the buggy behaviour this library has when there are more than one slide box in the page. Before this, selecting one tab of tabbedSlideBox would trigger the event to all the slide boxes in the page, resulting in an unwanted sync.
If you want to be sure that your slide box will be handled as one, you can add the delegate-handle="infoHandler" attribute to the <ion-slide-box> tag.